Semi-Working restoration
Restoration works, however no restoration is done if not all blocks
within one square are old. This is caused by a bug in CoreProtect where
restore() does not change the "isRolledBack"-state, which means that
these will be skipped later. This can be fixed within the plugin by not
checking whether the block is rolled back or not, but this is
sub-optimal and fixes will therefore wait till the CoreProtect team has
responded to this bug.