Alpha release
First functional version of the plugin. Regestoration tasks are working
as expected and restoring everything outside claims.

Build is relatively stable: tested in a world with few chunks generated,
with a restore radius of 100 000 blocks. Regeneration task finished
successfully after a few minutes.

The plugin in it's current state will can regenerate one block-row
within claims if claim-padding is set to zero.