Config and config version validation
The plugin is now able to validate the config file, find errors and give
feedback to consle. If an invalid version is found, the config is backed
up, and a new one is generated for the current version.

The plugin will now recognize outdated and newer version of the config.
Next up is importing settings from old config to new config.